requirements.txt
streamlit
psycopg2-binary
pandas
pyvis
networkx
requests
confluent_kafka
streamlit-autorefresh
